refactor(desktop): isolate Windows shell integration
